refactor: use middleware
