+++ next.config.js
@@ -3,6 +3,7 @@ module.exports = {
   reactStrictMode: true,
   experimental: {
     appDir: true,
+    allowMiddlewareResponseBody: true,
   },
   headers() {
     return [
