+++ src/seo.ts
@@ -10,5 +10,30 @@ const keywords: string = [
   "Самара",
   "Web",
 ].join(", ");
+const getSiteMap = (host: string): string =>
+  `
+  <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+<url>
+<loc>https://${host}/</loc>
+</url>
+</urlset>
+  `;
+const getRobotsTxt = (hostName: string, isAllow: boolean): string => {
+  const robots: Record<string, string> = isAllow
+    ? {
+        "User-agent": "*",
+        Allow: "/",
+        Sitemap: `https://${hostName}/sitemap.xml`,
+      }
+    : {
+        "User-agent": "*",
+        Disallow: "/",
+        Sitemap: `https://${hostName}/sitemap.xml`,
+      };
 
-export { title, description, keywords };
+  return Object.keys(robots)
+    .map((i) => `${i}: ${robots[i]}`)
+    .join("\n");
+};
+
+export { title, description, keywords, getSiteMap, getRobotsTxt };
